Refactor stitching
