+++ PythonRegression/tests/features/machine3/3_blowball_tests.feature
@@ -1,7 +1,7 @@
 Feature: Test GTTA for blowballs
 
 	Scenario: GTTA is called 1000 times on nodeA
-		GTTA should be called 1000 times on a node. The returned 
+		GTTA should be called 1000 times on a node. The returned
 		transactions will be checked to find the number of referenced 
 		milestones. The total percentage of returned milestones should 
 		be less than 5%.
