+++ PythonRegression/util/threading_logic/pool_logic.py
@@ -45,3 +45,4 @@ def fetch_results(future_result, timeout):
         return response
     except Exception as err:
         logger.debug(err)
+        logger.info(err)
